Implemented My workouts page, Edit/Add workout, Suggested workouts, updated routes
Created html files for the My Workouts page along with the Suggested Workouts which contain videos of Leg day, Upper Body day, and Core day workouts each in their separate html file. Also implemented the Add a Workout html page and the Edit a Workout html page. Updated the routes page to create routes for each html file and updated the dashboard along with the base.html to reflect changes. Next up is the database.